Rename isAintCase to isAncillaryInterimCase
